Add 10.6 to 11.1 upgrade to mysql_upgrade test
With the upcoming release of MariaDB 11 series, it is important to add
proper upgrade coverage from 10.6 to 11.1.

The 10.6 dump file comes from MariaDB version 10.6.3, the GA release of
the 10.6 series.

All new code of the whole pull request, including one or several files
that are either new files or modified ones, are contributed under the
BSD-new license. I am contributing on behalf of my employer Amazon Web
Services, Inc.